test: new structure test fixed
